+++ src/app/register/store/info/page.tsx
@@ -12,7 +12,7 @@ import TextField from '@/components/ui/text-field';
 import { usePostStoreRegistration } from '@/hooks/query/store/usePostStoreRegistration';
 import { phoneRegex } from '@/utils/regex';
 import useSearchAddress from '@/hooks/useSearchAddress';
-import ImageUploader from '../(components)/image-uploader';
+import ImageUploader from '@/components/image-uploader';
 import BankSelectSheet from './_components/bank-select-sheet';
 
 type StoreForm = z.infer<typeof schema>;
